Move SelectDialogFactory closer to where it is being used
